+++ tomcat/src/main/java/nextstep/jwp/service/LoginService.java
@@ -3,17 +3,16 @@ package nextstep.jwp.service;
 import java.util.Map;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.AuthenticationException;
-import nextstep.jwp.exception.DuplicationException;
 import nextstep.jwp.exception.NotFoundException;
 import nextstep.jwp.model.User;
 import org.apache.catalina.session.Session;
 import org.apache.catalina.session.SessionManager;
 
-public class Service {
+public class LoginService {
 
     private final SessionManager sessionManager;
 
-    public Service() {
+    public LoginService() {
         this.sessionManager = new SessionManager();
     }
 
@@ -42,17 +41,5 @@ public class Service {
 
     public boolean isAlreadyLogin(final Session session) {
         return sessionManager.findSession(session.getId()) != null;
-    }
-
-    public void register(final Map<String, String> parameters) {
-        final var account = parameters.get("account");
-        final var password = parameters.get("password");
-        final var email = parameters.get("email");
-
-        if (InMemoryUserRepository.findByAccount(account).isPresent()) {
-            throw new DuplicationException("이미 존재하는 아이디입니다.");
-        }
-
-        InMemoryUserRepository.save(new User(account, password, email));
     }
 }
